+++ js/ui/menu/ui.menu.js
@@ -6,7 +6,7 @@ var $ = require("../../core/renderer"),
     each = require("../../core/utils/iterator").each,
     typeUtils = require("../../core/utils/type"),
     extend = require("../../core/utils/extend").extend,
-    getElementMaxHeightByWindow = require("../../core/utils/dom").getElementMaxHeightByWindow,
+    getElementMaxHeightByWindow = require("../overlay/utils").getElementMaxHeightByWindow,
     eventUtils = require("../../events/utils"),
     pointerEvents = require("../../events/pointer"),
     hoverEvents = require("../../events/hover"),
